dont leak hash functions
